+++ example/lib/example5_themes_playground/widgets/panels/app_bar_settings/app_bar_settings.dart
@@ -276,6 +276,9 @@ class AppBarSettings extends StatelessWidget {
                   }
                 : null,
           ),
+        const Divider(),
+        const BottomAppBarShowcase(),
+        const SizedBox(height: 16),
       ],
     );
   }